Rename DLinkMotionSensorConfig -> DLinkThingConfig
